ME-199: Added base class for testing resources
Updated __all__ indices
