+++ mfit/resources/tests/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
+from mfit.resources.tests.base import RootBase
 from mfit.resources.tests.base import Base
 
-__all__ = ['Base']
+__all__ = ['Base', 'RootBase']
